Model data for store
Allow users to create a new store. A store has preference, and currently
those preference are a preferred language, currency, and country for the
store. Users have a 1:1 relationship with a store, and they also have a
relationship with a preferred language, currency, and country.

At some point it might be appropriate to create a user preferences
table.